food_detection_model/requirements.txt
torch
opencv-python
flask
streamlit
requests
numpy
pillow
